fix(dal): fix asset upgrade handling of variant differences
The copy/paste logic cannot be used when upgrading an asset to a new
schema variant, since it assumes the two components are on the same
variant. Instead, just set the root attribute value with the view from
the old component, and reset prototypes to the default if needed. This
will ensure we are graceful about prop differences between the two
variants, and we will create map and array elements correctly (if the
map/array exists in the new variant).

Also fixed by this pr:

  1. some attribute functions were dropped on upgrade if they took
     "code", "qualification", or the "resource" props as inputs (or a
     child prop of those).
  2. we were not correctly generating the index or key entry when
     debugging components, which had implications for copy and paste
     correctness.