Include Components only in the base change set that would be actively removed by a change set when building the summary diagram
We now look at the series of updates that would be applied if we were to
merge a change set into its base, and include "virtual" summary diagram
component entries for any Components that would be removed and that only
exist in the base change set. These virtual entries will have
`from_base_change_set` set to true to signal that they don't actually
exist in the snapshot for the change set.