Make Update::ReplaceSubgraph only import a single node, instead of the entire subgraph
The problem with importing the entire subgraph is that we also still
generate `Update` entries for things below the root of the
`Update::ReplaceSubgraph`, and that `find_in_self_or_create_using_onto`
+ `replace_references` ends up with the uinion of the "child" nodes from
both `self` and `onto` as the child nodes in the resulting graph.

What we want is either only the child nodes from `onto` with no further
`Update` entries for things below the root of the
`Update::ReplaceSubgraph`, or for only the specific node referenced in
`Update::ReplaceSubgraph` to be replaced, to maintain the pre-existing
child nodes in `self`, and to let the `Update` entries for things below
the root specified in `Update::ReplaceSubgraph` to handle the rest of
the subgraph update/import.