Add regression test for existing code gen funcs and new variants
This commit adds a regression test for attaching existing code gen funcs
to new schema variants. Specifically, the "attaching" step works, but
when the "Secrets" input is added, sometimes we are unable to regenerate
the variant due the following error: "pkg error: Can't convert identity
to LeafInputLocation".

The bug appears to no longer exist on main, but the test should still be
valuable in catching a future regression.

Secondary changes:
- Rename "find_by_name" to "find_id_by_name" when looking for funcs by
  name

Signed-off-by: Nick Gerace <nick@systeminit.com>