Centralize and modernize all documentation
This commit centralizes and modernizes all documentation. The goal for
this centralized document is to provide a nexus for notes and
long-living documentation for using and developing the System Initiative
software.

The centralized document does not strive to contain feature-specific
documentation as the product is ever changing. It is simply designed as
a hub for bucketed information that needs a home (e.g. "do not run SI
on a toaster, please").

The new documentation format takes advantage of GitHub markdown
rendering.

All section and pathing lookups have been modified as a result of these
changes.

Signed-off-by: Nick Gerace <nick@systeminit.com>