Show all funcs for all versions of an asset
This commit contains a bunch of changes related to authoring schemas
(assets), schema variants (asset versions) and funcs involved when
authoring them. Primarily, this commit is focused on showing all funcs
for all versions of an asset to ensure the user has all the data they
need for authoring.

When you regenerate an asset when there are components on the diagram
using it, a new version of the asset is created (a new schema variant).
As a result, the default schema variant changes for the asset (schema).
If you were to edit the output location or change anything that would
affect the bindings the asset, upon renegeration, you could lose access
to the func.

Here's what we do to solve that problem, among others that are
downstream to that problem:
- List ALL funcs in existence when asked, rather than just those that
  belong to assets (i.e. allow users to attach funcs that are free
  floating and aren't attached to any assets)
- List all funcs for all versions of an asset in the "ASSET FUNCTIONS"
  section (i.e. if a component is using an older version of an asset
  and that older version is using a func that is no longer in use, you
  can now access that func)
- Do not clone funcs on regeneration since the user should be explicit
  about changing the func to something else

Signed-off-by: Nick Gerace <nick@systeminit.com>