fix(dal): pick correct ordering node in ReplaceSubgraph
If an ordered node's children are replaced, and the RemoveEdge
operations come before the ReplaceSubgraph update for the ordering node,
the ordering node's vector clock write will have been incremented
*during the update*, which leads us to pick the wrong node, with the
result that the ordering node now has incorrect children in its order
property. This fixes that by not incrementing the vector clock write for
the ordering node when processing a RemoveEdge update (and *only* when
processing a RemoveEdge update).