Retain validations from asset func on regeneration
This commit retains validations from asset funcs upon regeneration.
Currently, we successfully execute the asset func when regenerating
an asset. We see the validations in the resulting definition. However,
when exporting using the definition and converting it into a spec, the
validations are dropped. This commit ensures that are no longer dropped.

This commit also contains minor changes to how validations appear in the
attribute panel. The existing classes relied on colors outside of our
asset panel and neither matcher our light nor our dark theme. While the
current colors are not likely the final ones, they are closer to our
design language (even if only a little).

Disclaimer: this will require all affected assets to be regenerated and
re-exported, as applicable.

Signed-off-by: Nick Gerace <nick@systeminit.com>