Restore Component->Component connections from base changeset whenever possible
Rather than creating a new API endpoint & method for attempting to
restore a `Component -> Component` (`Output Socket -> Input Socket`)
connection, `Component::connect` (and by extension
`diagram/create_connection`) will now check the base change set to see
if an equivalent connection exists there. If it does, then we import
that pre-existing connection into the current change set. If it does
not, then we'll create the connection as a brand new one.

This means that if you remove a connection, decide you didn’t actually
want to do that, and re-draw it instead of clicking the “restore
connection” button, we’ll do the right thing and restore it, so you
don’t end up seeing a remove+add of the “same” connection, and should
generally help reduce the potential churn in looking at diffs.