fix(dal): export default values correctly
A bit of code left over from the old system (where default values were
the attribute value for a prop at the schema variant context) was
grabbing the first attribute value and making it the default for props
if a component existed on export. This fixes it and fixes the export of
default values for strings.

Default values now can only be defined in the asset spec.

Co-Authored-By: Paul Stack <paul@systeminit.com>