feat(web): typescript linter should chill out
This makes our typescript configuration a whole lot less likely to throw
errors. It makes it a bit more straightforward to write code that we
know works but might be a bit YOLO.

It also adds an 'any' type for lodash, while we figure out how to get
the actual types in.

Signed-off-by: Adam Jacob <adam@systeminit.com>