Remove remaining old test builtins
This commit removes the remaining old test builtins. With this change,
we no longer need our "prelude-si" targets to track "pkgs" or
"pkgs_path".

As a result of these changes, two "fake" builtins were added to replace
the old pkgs. Ideally, these builtins would not resemble real world
resources in order to reduce confusion, but given the complexity of
fixing tests in place with as little interference as possible, this
commit opts to keep the existing structures of each asset in place.

Signed-off-by: Nick Gerace <nick@systeminit.com>