feat(dal): Auto migrate snapshots.
On SDF boot, attempt to automatically migrate all snapshots for a
deployment, beginning with the builtin workspace's snapshot. Follows the
"based_on_change_set_id" paths, treating the snapshots as a dependency
graph, so that shared clock ids are migrated to the new clock ids
correctly.

Once this code is deployed, SDF will panic if it encounters a 'legacy'
snapshot.

Co-Authored-By: Jacob Helwig <jacob@systeminit.com>