how that assets can be updated
This commit shows that an asset can be updated. This functionality is
achieved by asking the module index what the latest module is for the
corresponding local hash.

Why not use the SchemaId instead of the local hash? Existing SI
deployments may not have the SchemaId in their local modules since
the module-to-schema enforcement and corresponding table columns were
only put in place recently.

One consideration for the future: we may want the "can I update?" logic
to be moved into a different route in order to not hold up listing
variants. For now, the determination logic is wrapper in an infallible
wrapper in case something goes wrong, but that does not prevent poor
performance.

Secondary changes:
- Ensure "app/web" no longer mutates the "canUpdate" field for
  SchemaVariants
- Add README for `module-index` with instructions for local development
- Remove SchemaVariantListEntry as it is no longer needed

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Zachary Hamm <zack@systeminit.com>