feat(si-data-pg): Discard query plans when recycling connections
If a table's structure changes, cached query plans against that table
need to be invalidated, or postgresql will return an error. This change
prevents that error after migrating the database in a production system
running pb_bouncer, which holds on to connections and reuses them even
if our services are restarted. We could avoid needing to discard plans
by selecting exactly the columns we need instead of SELECT * (unless
the column type changes!)

This issue never hit us before because we haven't changed table
structures much since adding pg_bouncer to the stack.

Co-Authored-By: Jacob Helwig <jacob@systeminit.com>