Indicate what modules can be installed and updated
This commit provides the functionality to be able to know what modules
can be installed and updated. In the last PR (#4124), we were able to
indicate what assets could be updated, but it came with two caveats:

1. The logic for communicating with the module index was wrapped inside
   the route to list all user facing schema variants. While the logic
   was wrapped in an infallible wrapper, poor performance could still
   occur.
2. It did not convey what modules could be installed.

This commit moves and refactors the existing logic into a new route. In
addition, providing specific hashes is no longer required for the
request.

Disclaimer:

Upgradeable modules are displayed in the UI, but installable modules
are not. However, the backend route returns both upgradeable and
installable modules.

Secondary changes:

- Move logic for listing user facing schema variants into the dal (can
  be tested in integration tests and used by multiple routes)
- Add regression test for listing user facing schema variants
- Split module index types into its own crate (re-exported by the client
  crate for users of the client)
- Add module syncing types to frontend types crate (using the new module
  index types crate as well)
- Split v2 variant route business logic into sub-modules
- Fix layer cache crate cargo warning for specifying dev dependencies
- Remove module index client dependency from dal-test

Additional UI changes:
- Load modules on the front end, and use that data for canUpdate
- Add refresh button for syncing modules on demand in the asset list
- Make the "thank you" message for contributing cleaner

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: John Obelenus <jobelenus@systeminit.com>