Increase log level and log when module not found for schema
This commit increases the log level for module syncing as well as tunes
inner logs' levels. In addition, we now log instead of returning an
error because either the deployment existed during a time where schemas
were not one-to-one with modules, or the developer is testing schemas
without corresponding modules (e.g. recent AWS IAM creation efforts).

Signed-off-by: Nick Gerace <nick@systeminit.com>