Add ability to update assets (upgrade modules)
This commit adds the ability to update assets, which is known on the
backend as "upgrading modules". There are two ways to do so:

1. Update all assets at once by clicking the update button on the upper
   left-hand side of the asset list panel
2. Update individual assets by selecting them and then clicking the
   update button on the upper right-hand side of the asset card on the
   right-hand side of the application

In both cases, we unset the currently selected asset in order to avoid
a mismatched state. In the future, we may be able to avoid unsetting
(only for the "update all" case) by checking if the currently selected
asset is not in the "upgradeable" list.

This commit also ensures that unlocked schema variants do not appear
as "upgradeable".

Signed-off-by: Nick Gerace <nick@systeminit.com>