Initial API to rebase a Change Set onto its base Change Set
This has an identical interface to the API for applying a Change Set to
its base Change Set, but the result is to rebase in the opposite
direction. By rebasing the Change Set onto the base Change Set, we'll
incorporate all of the changes that we have not yet seen from the base
Change Set.

This does not yet handle re-running functions (enqueuing
DependentValuesUpdate) for things affected by the rebase.