feat(dal, sdf): Migrate workspace at create time
When a workspace was created, we used the Default workspace as the starting point of it's existence. This meant that if the default workspace got out of sync with the real world, then each new workspace created would also be out of sync

We now want to replace the login in workspace::new with the migration of the workspace and not even care about the default workspace. This will allow new users to continually get the latest builtins without the need to tell them to migrate all of the builtins on installation