feat(auth-portal): Redirect to the default workspace on login
When a user logs into the app (not from a redirect), we want to take the user to their default workspace.
The default workspace is chosen in the following order:
* If the workspace has a flag of isDefault
* The first production workspace owned by the user

This means that by shipping this code, we only affect users who have a production workspace. This will not redirect users with a local workspace