Ensure recycled PG connections do not have lingering open transactions
In addition to the in-process pool of PG connections, we also have
connection pooling & recycling happening in pgbouncer. Because there are
multiple layers that might result in us using an already established
PG connection that is an unknown state, we need to make sure that we're
also not dealing with a lingering open transaction in addition to the
things we're already doing to make sure the connection is "clean" before
we start using it.