Temporarily silence PG warnings during connection setup
In the case where we get back a clean (or at least cleanly committed)
connection from either the in-process PG connection pool, or from
pgbouncer, doing `ROLLBACK` will issue a warning that shows up in the
logs. By temporarily setting the minimum log level to error before we do
the `ROLLBACK`, and restoring it to the default value after rolling back
we should no longer see the warning about not being in a transaction,
while still getting other warnings.