Add configurable timeout to function execution in lang-js
This commit adds timeouts to lang-js function execution as well as the
ability to configure them. By default, functions can run for a maximum
of 30 minutes. This is tunable by a command-line argument, which can be
used by cyclone instances. Both cyclone and veritech have tunables
for the lang server's function timeout.

In addition to these changes, dead example code has been removed and
the docs for lang-js have been revitalized, and the test suite timeout
has been increased for lang-js unit tests. Not only that, but we now
"process.exit(0)" upon successful lang-js runs, which you can read the
comment in the source code to learn more about.

For testing, a new unit test has been added that uses a 1 second timeout
in the lang-js test suite.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: John Obelenus <jobelenus@systeminit.com>