Retry removing action nodes upon completion
This commit retries removing actions nodes upon action completion. This
is a workaround for another bug where a "RemoveEdge" update is not
being detected for children of the action category node when the
"to_rebase" snapshot address has changed in between commit time in the
job and the time that the rebaser performs update and conflict
detection.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Brit Myers <brit@systeminit.com>