feat(dal): correct transforms to ordering nodes on rebase
This adds a framework designed by @jhelwig for us to detect when graph
merges will create nonsensical results, and smooth them out. When we
receive a set of OTs (updates) to apply to our branch, we ask the
relevant nodes in our graph whether the set of updates makes sense, and
give them a chance to massage the updates (modify, add, remove, etc.)
until it generates a consistent graph.

The CorrectTransforms trait, implemented by all NodeWeights, is where
the code for correcting transforms/updates goes. It is passed the entire
set of updates and must return the set of updates (if there are no
changes, it must return the updates unchanged).

Right now, the only conflict we're resolving is the Ordering node, which
determines the order of values in an array value. If two people add
ports to a docker image's ExposedPorts in a short time, this code will
try to honor both of their choices, producing an array with both ports
added.

This PR also adds the expected module for DAL testing, which eliminates
a lot of the boilerplate expect calls in our tests. Most of these calls
aren't the actual thing we're trying to test, and they took up a lot of
extra horizontal and vertical space (often on their own line). At least
for me, the noise was high enough that it made it harder to figure out
what the test was actually testing :) In the tests I added, expected
stuff cut the lines of code per test by more than half (probably 2/3
or so at a rough estimate).

Co-Authored-By: John Keiser <john@johnkeiser.com>
Co-Authored-By: Fletcher Nichol <fletcher@systeminit.com>
Co-Authored-By: Jacob Helwig <jacob@systeminit.com>