Add cyclone client timeout and make func execution timeouts configurable
This commit adds a timeout wrapper when executing functions via the
cyclone client in veritech. In addition, the default timeout value can
be overriden as well as the recently added timeout value for cyclone
server's lang server process.

As a result of this commit and recent commits, the following values
are now configurable:

- 1) cyclone client timeout
  - dynamically configurable via veritech args and/or config file
- 2) cyclone server timeout
  - statically configurable with cyclone args and/or config file
- 3) lang-js function timeout
  - statically configurable with cyclone args and/or config file
  - dynamically configurable with lang-js args, but they are also
    statically configured due to how cyclone is invoked via firecracker
    (see the rootfs builder script for more information)

This commit also renames some functions for clarity, fixes some doc
comments and does some cleanup.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Victor Bustamante <victor@systeminit.com>