Fix function timeout set to 0 by default
This is an embarrassing one. We tested #4318 extensively, but due to a
last minute change with appeasing the test harness config, the default
value of the veritech config would be "0" unless you specifically
called the right constructor method to provide the correct value.

As a result of this change, the config for veritech is more clear anyway
because the config option should be... well, optional. So now, whenever
the config does not specify and option, we will fallback to the 35
minute timeout. Sorry, Victor, this is on me.

Signed-off-by: Nick Gerace <nick@systeminit.com>