Cleanup and encapsulate the logic for finding inferred connections
This PR includes bug fixes and cleanup as the functionality of Frames has evolved since it was first ported to the new engine and there was lingering traversal bugs depending on if you were searching for Outgoing or Incoming inferred connections. 

There is a new module InferredConnectionGraph which encapsulates the logic around finding Inferred Connections (incoming and outgoing), moving this logic out of the Component module.  

All of the existing tests apply and I added a new one to account for the fix to ensure that multiple Input Sockets for an Up Frame don't connect to the same Output Socket.