fix(rebaser): resolve timing window overlaps & add more resiliency
This change makes several timing interval adjustments which likely have
led to the "2 leaders" issue as seen in larger deployments.
Additionally, there are several fixes for pessimistic scenarios where
multiple debouncer tasks believe themselves to be the leader for the
same change set.

Timing Window Overlaps
----------------------

There are 3 timings here to consider:

1. the max age (that is, a "ttl") of a key in this K/V bucket
2. the "keepalive" interval (how frequent a leader updates the key to
   reset the key's ttl)
3. the "check for missing key" interval (how often to check whether or
   not a key is not present in the store)

Before these changes, if the max age of a key would be 5 seconds, then
the keepalive interval was 5 * 0.85 = 4.25 seconds and the missing key
interval was 5 seconds (hard-coded).

If, due to either high load on the system running the Tokio reactors or
due to high network load to the NATS server which might delay writes,
the keepalive interval was to trip over the ttl max age, then the NATS
server would (correctly) expire the key and any "waiting to be leader"
actors could attempt to acquire the key. Meanwhile the leader has "lost"
ownership of the key--and doesn't realize it. So this leader is now not
the leader (even though it thinks it is) and another "waiting to be
leader" has now (correctly) become the leader.

The issue is that now we have 2 "leaders" that are calling `kv.update()`
providing their last write revision number. This is by design as a
mechanism to detect other writers who should *not* be present. In the
above scenario however, there *are* other rogue writers.

There are 2 halves to this issue: reducing the chances that a timing
window overlap occurs and better detecting the scenario so that all
actors can work to resolve the conflict.

The window timing fix here is this:

- Increase the "check for missing key" interval to be larger than the
  key max age. Before this change the 2 values were exactly the same.
  This change updates the check interval to `max_age * 1.5`, so 7.5
  seconds in this case.
- Decrease the "keepalive" interval to be well inside the key max age
  duration. Before this change the value was `max_age * 0.85` and now it
  is `max_age * 0.5`, so 2.5 seconds in this case.

The detection and correction half are addressed as part of the "adding
more resiliency" section below.

Adding More Resiliency
----------------------

Several updates are included so that the `DvuDebouncerTask` can continue
to operate even in the face of a potential "2 leaders" scenario that is
beginning to happen.

- When the "check for missing key" interval fires, first check if the
  key does not exist (via `kv.entry(key).is_none()`) before attempting
  to acquire the key. This helps to distinguish between "a key was never
  before present" and "a key was expired by the server" with "a key was
  deleted/purged by another actor".
- A oneshot message originating from a keepalive task is checked for by
  the debouncer task while in leader mode. If the keepalive task fails
  to update its key (due to another writer, a key just aging out, etc.),
  then it can immediately signal the debouncer task. The debouncer task
  will stop its "wait-and-check-on-dvu" loop, immediately flip its state
  to "waiting" and leave the key undisturbed. It's important to not
  attempt any further writes/deletes/etc to let the key age out. When in
  doubt, any actors believing themselves to be leaders, should just back
  away...
- Immediately before a leader enqueues a dvu job it attempts to update
  the key's value status to "running". Now, if this task fails, then the
  leader will return out of the leader loop, immediately flip its state
  to "waiting" and leave the key undisturbed as above. In this way we
  hope to avoid kicking off a dvu while another might be running (or in
  the process of being kicked off).
- When the leader has successfully performed a dvu job it will demote
  itself by purging the change set key, allowing other rebasers to
  observe this demotion and try to acquire the lock for the next
  iteration. Before this change, when a `kv.purge_expect_revision()`
  call failed, the code attempted to resolve the situation by calling
  `kv.purge` (without a revision). This was incorrect behavior as a
  delete/purge will leave a tombstone entry and increase the `revision`
  number for that key. With this change, if the key can't be purged
  using its last known revision number, it simply logs that fact,
  finishes its leader demotion (by flipping to the "waiting" state) and
  does nothing further to the key (for the same reasons as above).

Welcome to my novel, I hope you enjoyed it ;)

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>