feat(dal): dvus for secret and attribute changes from another change set
When a secret changes or an attribute value changes, and those changes
are replayed from head onto another change set, we need to add these
values into the dvu roots so that any values that depend on them are
recomputed in the context of that change set.

This is a pretty big hammer: if we detect any change in an attribute
on replay that is already present in the graph, we add it to the dvu
roots. This potentially means adding quite a few roots to the graph on
big replays, and it also means recalculating a lot of values
that really do not need to be recalculated.

Right now we don't have enough information in the transform corrections
to be more surgical. What we would need would be

  1. The snapshot of the graph that we are replaying updates *from*,
     after it has been updated.
  2. Input socket kind on the Socket node weight.

This would allow us to find all the components that exist only the
change set that is receiving the replay and add *just* their input
sockets to the DVU, which would mean we recompute just the components
that were not already computed, and not recompute the ones that are
already properly computed when the DVU ran on the change set that
originally made these changes.