fix(rebaser): when error occurs, abandon leadership & don't purge key
This change adds one more internal state to the dependent values update
task "debouncer" task when running as leader:
`DebouncerState::AbandoningLeadership`. When a key keepalive write
fails, the returned state is now `AbandoningLeadership`, ensuring that a
final purge of the key will not be attempted (remember that a purge or
delete is another write to the key which will cause a revision update).

With this logic fix, when a leader notices write corruption to its key,
it will now immediately stop writing to the key, transition back into
`WaitingToBecomeLeader` and thus should allow the key to age out
(assuming no other writers are active).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>