Accept schema variant id when cloning
The front end is sending the schema variant id, and as we were using the schema id to get the variant anyways, this change modifies the sdf route to accept the variant id.  This also enables users to clone any variant they can view (including unlocked). This change also fixes up the asset name picker and adds "-Clone" to the end of the display name so it's differentiated from what you were cloning.